+++ CHANGELOG.md
@@ -6,6 +6,9 @@
 - Support for .NET 6.0
 - User-Agent header to http requests
 - Business events capturing API `SendBizEvent`
+- `Session.ReportNetworkTechnology(string technology)`
+- `Session.ReportCarrier(string carrier)`
+- `Session.ReportConnectionType(ConnectionType connectionType)`
 
 ### Removed
 - Support for .NET 4.5.1 / 4.5 / 4.0